Setup/pip_requirements.txt
mysql-connector-python
pysondb